Refactoring: restructure http-server
