+++ lib/http-server/src/http_response.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 use pheidippides_utils::utils::CaseInsensitiveString;
-use super::Header;
+use pheidippides_utils::utils::Header;
 
 pub struct HttpResponse(Vec<u8>);
